# GestionHospitalAPI/OpenAPi.yaml
openapi: 3.0.0
info:
  title: Hospital Management API
  version: 1.0.0
  description: API para la gestión de un hospital incluyendo pacientes, camas, habitaciones, asignaciones, historial de altas, usuarios y roles.
servers:
  - url: http://localhost:5000/api
paths:
  /pacientes:
    get:
      summary: Obtener todos los pacientes
      responses:
        "200":
          description: Lista de pacientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Paciente"
    post:
      summary: Crear un nuevo paciente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Paciente"
      responses:
        "201":
          description: Paciente creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paciente"
  /pacientes/{ID_Paciente}:
    get:
      summary: Obtener un paciente específico
      parameters:
        - name: ID_Paciente
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detalles del paciente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paciente"
    put:
      summary: Editar un paciente existente
      parameters:
        - name: ID_Paciente
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Paciente"
      responses:
        "200":
          description: Paciente actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paciente"
    delete:
      summary: Eliminar un paciente existente
      parameters:
        - name: ID_Paciente
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Paciente eliminado
...
components:
  schemas:
    Paciente:
      type: object
      properties:
        ID_Paciente:
          type: integer
          example: 1
        Nombre:
          type: string
          example: "Juan Pérez"
        Edad:
          type: integer
          example: 30
        Sintomas:
          type: string
          example: "Dolor de cabeza"
        Estado:
          type: string
          example: "Pendiente de cama"
        Fecha_Registro:
          type: string
          format: date-time
          example: "2023-07-08T12:34:56Z"
        Seguridad_Social:
          type: string
          example: "123456789"
        Direccion:
          type: string
          example: "Calle Falsa 123"
        Telefono:
          type: string
          example: "555-1234"
        Email:
          type: string
          example: "juan.perez@example.com"
        Historial_Medico:
          type: string
          example: "Sin historial relevante"
...